Started the Serialization Kit.
This kit extends the functionality of the built in serialization functionality
of .NET. The new functionality includes features such as instance specific
serialization.

ExtendedFormatter wraps around another formatter and provides the ability to do
instance specific serialization.

UniqueObjectSerializer is a class that provides a layer of indirection and
actually is what is encoded into the instruction stream. This is done by having
unique object serializer act as an intermediary between the instance and the
formatting stream. The object requires a serialization action delegate and
target instance to instantiate. When this object is deserialized, the resulting
serialization information and streaming context are stored. The act of
deserialization occurs by calling Deserialize which takes in an object
instantiator function and DeserializationAction delegate which performs the act
of deserialization unique to this instance.